Expose /.well-known/jwks-okta for Okta API services type App
